-stable 1/16/2017

+++ core/src/com/mygdx/game/entity/map.java
@@ -33,9 +33,9 @@ public class map {
 
     public void setViewportHeight (float h){camera.viewportHeight = h;}
 
-    public float getMapWidth(){return map.getProperties().get("width",Integer.class)*map.getProperties().get("tilewidth",Integer.class);}
+    public float getMapWidth(){return (float) map.getProperties().get("width",Integer.class)*map.getProperties().get("tilewidth",Integer.class);}
 
-    public float getMapHeight(){return map.getProperties().get("height",Integer.class)*map.getProperties().get("tileheight",Integer.class);}
+    public float getMapHeight(){return (float) map.getProperties().get("height",Integer.class)*map.getProperties().get("tileheight",Integer.class);}
 
     public OrthographicCamera getCamera(){return  camera;}
